minor tweaks
I made a few tweaks, such as including fighting style for Fighter, though I'm now figuring out that part of how I'm organizing stuff is conflicting. Since `player_sheet['Features & Traits']` is a list with a dictionary for fighting style in it, I can't access the fighting style. It's probably not going to be too much of a pain to fix, but I'm going to take a break now.